Separate immutable 'object_values' from mutable 'entities' Apply some linting
