+++ app/tools.py
@@ -15,3 +15,9 @@ def get_formated_price(
         formated price (String)
     """
     return "{:0.0{}f}".format(amount, precision)
+
+
+def _strict_int_validator(cls, v):
+    if Decimal(v) != int(v):
+        raise ValueError
+    return v
